pyproject.toml
[project]
name = "kubek"
version = "0.4.2"
description = "kubectl extensions"
readme = "README.md"
authors = [
    {name = "ohmycoffe",email = "ohmycoffe1@gmail.com"}
]
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.11"
keywords = ["kubernetes", "kubectl", "kubectl plugins","cli", "devtools", "port forwarding", "dotenv"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "typer (>=0.25.1,<0.26.0)",
    "questionary (>=2.1.1,<3.0.0)",
    "psutil (>=7.2.2,<8.0.0)",
    "pydantic (>=2.13.4,<3.0.0)",
    "rich (>=15.0.0,<16.0.0)",
    "ruamel-yaml (>=0.18.0,<0.19.0)",

]

[project.scripts]
# Kubectl plugins must be prefixed with `kubectl-` and be discoverable in PATH
kubectl-envx = "envx.__main__:app"
kubectl-portfwd = "portfwd.__main__:app"

# Deprecated entries for backward compatibility, will be removed in the future
envx = "envx.__main__:deprecated_entry" # Deprecated, use `kubectl envx` instead

[project.urls]
Repository = "https://github.com/ohmycoffe/kubek"

[tool.poetry]
packages = [
    { from = "kubectl-envx/src", include = "envx" },
    { from = "kubectl-portfwd/src", include = "portfwd" },
    { from = "kubek-shared/src", include = "kubek" },
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.8"
pytest = "^9.0.3"
pre-commit = "^4.6.0"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
